+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - `Innmind\Immutable\Identity::lazy()`
 - `Innmind\Immutable\Identity::defer()`
 
+### Changed
+
+- `Innmind\Immutable\Sequence::toIdentity()` returns a lazy, deferred or in memory `Identity` based on the kind of `Sequence`
+
 ## 5.7.0 - 2024-06-25
 
 ### Added